- Updated user event to support checking for mute, username - Created functions for checking username, checking if user should be muted, and sending embed on join - Added string extension class - Created contains special chars & remove special chars extension functions - Refactoring to user event class - Updated database context for muted users table - Updated profile command to display roles & support database mute check - Updated mute, unmute, and auto unmute service to support database - Updated migrations

+++ Services/Rainbow Role.cs
@@ -31,7 +31,7 @@ public class RainbowRoleService : BackgroundService
                     SocketGuild? guildSocket = _client.GetGuild(guild.guildId);
                     if (guildSocket is not null)
                     {
-                        SocketRole? role = guildSocket.GetRole((ulong)guild.roleId);
+                        SocketRole? role = guildSocket.GetRole(guild.roleId);
                         if (role is not null)
                             await role.ModifyAsync(x => x.Color = Utilities.Miscallenous.RandomDiscordColour());
                     }
